requirements.txt
selenium
webdriver-manager
colorama
PyYaml
